Convert test file from mocha to jest
